Correction on quint32

+++ app/Network/TCPProtocolHelper.cpp
@@ -5,7 +5,7 @@
 // Login   <antoinegarcia@epitech.net>
 //
 // Started on  Tue Oct 27 02:58:48 2015 Antoine Garcia
-// Last update Sun Nov  1 08:31:11 2015 Antoine Garcia
+// Last update Sun Nov  1 17:41:42 2015 Nicolas Girardot
 //
 
 #include "TCPProtocolHelper.hh"
@@ -48,7 +48,7 @@ QByteArray	TCPProtocolHelper::createHandshake()
   QDataStream	out(&block, QIODevice::WriteOnly);
 
   out.setVersion(QDataStream::Qt_4_3);
-  out << quint8(1) << quint64(0) << quint16(str.size());
+  out << quint8(1) << quint32(0) << quint16(str.size());
   out.writeRawData(str.c_str(), str.size());
   return block;
 }
